test(webapp): limit dashboard watch fixture fan-out
